test: templating for human control tests
