+++ arm_control/test/human_control_tests.py
@@ -1,0 +1,1 @@
+from arm_control import human_arm_control
